+++ CHANGELOG.md
@@ -22,6 +22,9 @@ All changes in this project will be noted in this file.
 - `SkyResult` is now `SkyQueryResult`
 - The feature `async` is now `aio`
 - `Query::add` is now `Query::append` to avoid confusion with the `core::ops::Add` trait
+- `Element::Float` is now an `f32` instead of an `f64` because the Skyhash specification requires it
+  (this was mistakenly an `f64` but shouldn't be a problem because no actions returned any floating
+  point value)
 
 ## Version 0.6.2